Migrate to auto-zero metrics and avoid negative rates for GUI

+++ src/Modules/DigitizerNoiseMonitor/DigitizerNoiseMonitorModule.hpp
@@ -17,7 +17,7 @@ class DigitizerNoiseMonitorModule : public MonitorBaseModule {
   
   std::atomic<float> m_metric_pedestal[16];
   std::atomic<float> m_metric_rms[16];
-  std::atomic<double> m_metric_peaks[4][16];
+  std::atomic<float> m_metric_peaks[4][16];
 
  protected:
